# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aerospike Backup Service",
        "title": "Backup Service REST API Specification",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/backups/full": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available full backups.",
                "operationId": "getFullBackups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full backups by routine",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BackupDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/full/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available full backups for routine.",
                "operationId": "getFullBackupsForRoutine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full backups for routine",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BackupDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/incremental": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get available incremental backups.",
                "operationId": "getIncrementalBackups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incremental backups by routine",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BackupDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/incremental/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backup"
                ],
                "summary": "Get incremental backups for routine.",
                "operationId": "getIncrementalBackupsForRoutine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lower bound timestamp filter",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Upper bound timestamp filter",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incremental backups for routine",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BackupDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/backups/schedule/{name}": {
            "post": {
                "tags": [
                    "Backup"
                ],
                "summary": "Schedule a full backup once per routine name.",
                "operationId": "scheduleFullBackup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup routine name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Delay interval in milliseconds",
                        "name": "delay",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Returns the configuration for the service.",
                "operationId": "readConfig",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates the configuration for the service.",
                "operationId": "updateConfig",
                "parameters": [
                    {
                        "description": "config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/clusters": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all Aerospike clusters from the configuration.",
                "operationId": "readAllClusters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.AerospikeCluster"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/clusters/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads a specific Aerospike cluster from the configuration given its name.",
                "operationId": "readCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the Aerospike cluster",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    },
                    "404": {
                        "description": "The specified cluster could not be found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates an existing Aerospike cluster in the configuration.",
                "operationId": "updateCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "aerospike cluster",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Adds an Aerospike cluster to the config.",
                "operationId": "addCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "cluster info",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Deletes a cluster from the configuration by name.",
                "operationId": "deleteCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/policies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all policies from the configuration.",
                "operationId": "readPolicies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.BackupPolicy"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/policies/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads a backup policy from the configuration given its name.",
                "operationId": "readPolicy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the backup policy",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BackupPolicy"
                        }
                    },
                    "404": {
                        "description": "The specified policy could not be found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates an existing policy in the configuration.",
                "operationId": "updatePolicy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "policy name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "backup policy",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackupPolicy"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Adds a policy to the config.",
                "operationId": "addPolicy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "policy name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "backup policy",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackupPolicy"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Deletes a policy from the configuration by name.",
                "operationId": "deletePolicy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/routines": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all routines from the configuration.",
                "operationId": "readRoutines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.BackupRoutine"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/config/storage": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all storage from the configuration.",
                "operationId": "readAllStorage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.Storage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/restore/full": {
            "post": {
                "description": "Specify the directory parameter for the full backup restore.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Restore"
                ],
                "summary": "Trigger an asynchronous full restore operation.",
                "operationId": "restoreFull",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RestoreRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Config": {
            "description": "Config represents the service configuration file.",
            "type": "object",
            "properties": {
                "aerospike-clusters": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.AerospikeCluster"
                    }
                },
                "backup-policies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.BackupPolicy"
                    }
                },
                "backup-routines": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.BackupRoutine"
                    }
                },
                "secret-agent": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.SecretAgent"
                    }
                },
                "service": {
                    "$ref": "#/definitions/model.BackupServiceConfig"
                },
                "storage": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.Storage"
                    }
                }
            }
        },
...
    },
...
}